play.api.Plugin is deprecated. See https://www.playframework.com/documentation/2.4.x/PluginsToModules
