+++ play-2.4/swagger-play2/conf/application.conf
@@ -1,2 +1,4 @@
 # This is the main configuration file for the swagger-play2 plugin
 # ~~~~~
+
+play.modules.enabled  += "play.modules.swagger.SwaggerModule"
