--- play-2.4/swagger-play2/conf/play.plugins
@@ -1,1 +1,0 @@
-1000:play.modules.swagger.SwaggerPlugin
